pyproject.toml
[tool.poetry]
name = "Roboseiten"
version = "0.2.0"
description = ""
authors = ["J"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
selenium = "^4.14.0"
pyinstall = "^0.1.4"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
